backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0
cryptography>=3.0.0,<42.0.0
celery==5.3.4
python-dateutil==2.8.2
pandas==2.1.3
python-magic==0.4.27
email-validator==2.1.0
bcrypt==4.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
